Added plugins component
