+++ moninag/static/src/ts/app/settings/settings-help.component.html
@@ -1,0 +1,8 @@
+<div class="flex-container">
+    <label>If you have any questions, please contact us at:</label>
+    <a class="flex-container" fxLayout="row" fxLayoutAlign=" center" 
+    href="mailto:moninaginfo@gmail.com">
+        <md-icon>help</md-icon>
+        <span>moninaginfo@gmail.com</span>
+    </a>
+</div>
