+++ moninag/static/src/ts/app/settings/settings-help.component.ts
@@ -4,16 +4,7 @@ import { Component } from '@angular/core';
 @Component({
     selector: 'settings-help-app',
     styles: [require('../../../less/styles.less').toString()],
-    template: `
-        <label>If you have any questions, please contact us at:</label>
-        
-        <a class="flex-container" fxLayout="row" fxLayoutAlign=" center" 
-        href="mailto:moninaginfo@gmail.com">
-            <md-icon>help</md-icon>
-            <span> moninaginfo@gmail.com</span>
-        </a>
-        
-    `
+    template: require('./settings-help.component.html'),
 })
 
 export class SettingsHelpComponent { }
