+++ moninag/static/src/ts/app/settings/settings-menu.component.ts
@@ -1,14 +1,15 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 import { Router } from '@angular/router';
 import { FlexLayoutModule } from "@angular/flex-layout";
-import { UserProfileComponent } from '../user-profile/user-profile.component';
-import { SettingsHelpComponent } from './settings-help.component';
 
+import { SettingsPluginsComponent } from './plugins.component';
+import { SettingsHelpComponent } from './settings-help.component';
+import { UserProfileComponent } from '../user-profile/user-profile.component';
 
 @Component({
     selector: 'settings-menu',
-    template: require('./settings-menu.component.html'),
     styles: [require('./settings.less').toString()],
+    template: require('./settings-menu.component.html'),
 })
 
 export class SettingsMenuComponent { }
