+++ moninag/static/src/ts/app/settings/settings.less
@@ -33,8 +33,12 @@ settings-app {
             margin-right: 10px;
         }
     }
-    
 }
+
 .settings-menu-container .md-tab-header {
     margin-bottom: 27px
+}
+
+.table-plugins {
+    width: 70%
 }
